+++ Basil/include/Renderer/Renderer2D.h
@@ -33,6 +33,7 @@ namespace Basil
 			static void drawQuad(const glm::mat4& transform, const glm::vec4& color, int entityID = -1);
 			static void drawQuad(const glm::mat4& transform, const Shared<Texture2D>& texture, float textureScale = 1.0f, const glm::vec4& tintColor = glm::vec4(1.0f), int entityID = -1);
 			static void drawSprite(const glm::mat4& transform, SpriteRendererComponent& src, int entityID);
+			static void drawCircle(const glm::mat4& transform, const glm::vec4& color, float thickness = 1.0f, float fade = 0.005f, int entityID = -1);
 			static void resetStats();
 			static Statistics getStats();
 	};