+++ Pesto/src/Panels/PropertiesPanel.cpp
@@ -202,6 +202,16 @@ namespace Basil
 				}
 			}
 
+			// Circle renderer menu item - add a Circle Renderer Component
+			if (!entity.hasComponent<CircleRendererComponent>())
+			{
+				if (ImGui::MenuItem("Circle Renderer"))
+				{
+					entity.addComponent<CircleRendererComponent>();
+					ImGui::CloseCurrentPopup();
+				}
+			}
+
 			// Rigidbody 2D menu item - add a Rigidbody 2D Component
 			if (!entity.hasComponent<Rigidbody2DComponent>())
 			{
@@ -327,6 +337,13 @@ namespace Basil
 			}
 
 			ImGui::DragFloat("Tiling Factor", &component.tilingFactor, 0.1f, 0.0f, 100.0f);
+		});
+
+		drawComponent<CircleRendererComponent>("Circle Renderer", entity, [](auto& component)
+		{
+			ImGui::ColorEdit4("Color", glm::value_ptr(component.color));
+			ImGui::DragFloat("Thickness", &component.thickness, 0.025f, 0.0f, 1.0f);
+			ImGui::DragFloat("Fade", &component.fade, 0.00025f, 0.0f, 1.0f);
 		});
 
 		drawComponent<Rigidbody2DComponent>("Rigidbody 2D", entity, [](auto& component)